requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docker==6.1.3
drf-spectacular==0.26.4
flake8==6.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==1.7.1
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.9.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.4
websocket-client==1.6.1
